# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: |
    SubNation Digital Subscriptions Platform API.

    ## Authentication (round-93 / 93-A8 F-1 — contract truth)

    Two session mechanisms, both documented as security schemes and
    attached per operation family:

    - **User routes** (auth, orders, wallet, cart, coupons/validate,
      loyalty, referrals, support, notifications): an `auth_token`
      httpOnly cookie (primary) OR an `Authorization: Bearer <token>`
      header (fallback).
    - **Admin routes** (/admin/*, /coupons/admin*, the copilot family):
      an `admin_token` httpOnly cookie (8h, issued by POST
      /admin/login) OR its Bearer equivalent. POST /admin/login itself
      is the credential endpoint and needs no prior session.

    ## CSRF gate (applies to every state-changing request)

    Every POST/PUT/PATCH/DELETE under /api — regardless of cookie or
    Bearer authentication — must carry an Origin (or Referer) header
    matching the deployment's allow-list (CSRF_ALLOWED_ORIGINS /
    APP_ORIGINS / APP_URL); otherwise the request is rejected with
    403 FORBIDDEN before the route runs. Requests with neither Origin
    nor Referer are also rejected. Exemptions:
    /api/auth/firebase/session, /api/auth/firebase/refresh, /api/cwv,
    /api/webhook/*, and /health. Non-browser clients must therefore
    send a browser-like Origin header on mutations.

    ## Deployment-conditional behavior

...
servers:
  - url: /api
...
security:
  - cookieAuth: []
    bearerAuth: []
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: List users (requireAdmin + users scope; paginated)
      parameters:
        - name: search
          in: query
          description: Phone substring filter (LIKE %search%).
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          description: |
            1-based page number (offset = (page-1)*limit).
            Round-93 (A8 F-2): this parameter was documented but never
            implemented until this round — every request used to return
            the identical first 100 rows.
          schema:
            type: ["integer", "null"]
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Page size, clamped to 1..200 (default 100).
          schema:
            type: ["integer", "null"]
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing users permission scope (requirePermission guard)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: Edit a user — wallet adjustment/set, loyalty fields (requireAdmin + users scope; moves wallet money)
      description: |
        THE admin money-edit route. Wallet mutations
        (`wallet_adjustment` relative delta, or `wallet_balance` absolute
        set — send at most one of the two; `wallet_adjustment` wins if
        both are present) route through AdjustmentService: transactional
        optimistic-lock update + a wallet_ledger row of type=adjustment,
        so every monetary change stays reconstructable from the ledger
        (Constitution Principle I). Loyalty fields
        (`loyalty_points`, `loyalty_tier`) use a legacy direct-update
        path with no ledger.

        Mounted behind the idempotency middleware (routeKey
        admin.users.patch) so a network retry / admin double-click
        cannot double-credit — send a fresh Idempotency-Key per edit.
        The audit row records the fields changed and the resulting
        wallet balance.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Idempotency key (≥ 8 chars to engage; 24h dedup per admin +
            route). Currently advisory — absence logs a warning and the
            call proceeds; same key + different body → 409
            IDEMPOTENCY_KEY_REUSE, same key in flight → 409
            IDEMPOTENCY_IN_FLIGHT, replays carry Idempotent-Replayed /
            Idempotent-Original-At response headers.
          schema:
            type: string
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserBody"
      responses:
        "200":
          description: Updated user (wallet/loyalty current state)
          headers:
            Idempotent-Replayed:
              description: |
                "true" when this body is a cached replay of the original
                request (same Idempotency-Key + body).
              schema:
                type: string
                enum: ["true"]
            Idempotent-Original-At:
              description: |
                ISO-8601 completion timestamp of the original request —
                present on idempotent replays.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserUpdateResult"
        "400":
          description: |
            Invalid (non-integer) id, no editable field supplied, an
            out-of-bounds loyalty_points value (integer 0..10,000,000),
            or a rejected wallet amount (INVALID_AMOUNT / ZERO_DELTA /
            NEGATIVE_BALANCE from AdjustmentService).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin session missing/invalid (requireAdmin guard)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing users permission scope (requirePermission guard)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: |
            Wallet balance (or loyalty fields) changed concurrently —
            optimistic-lock race; the body code is CONFLICT (round-93
            A8 F-7: previously mislabeled INVALID_DATA) — retry the
            read-then-edit. Idempotency-key conflicts (same key re-used
            with a different body / same key still in flight — those
            bodies carry success:false and a legacy `message` alias
            alongside `error`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/IdempotencyConflictResponse"
...
components:
  securitySchemes:
    # Round-93 (93-A8 F-1): the spec previously documented zero security
    # definitions while the middleware reality is cookie-first with a
    # Bearer fallback on user routes, and a separate admin cookie (8h)
    # plus Bearer on admin routes. Both alternatives are expressed per
    # operation via `security: [{ schemeA: [], schemeB: [] }]` (same
    # object = OR).
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: |
        User session — httpOnly cookie set by the auth routes (Google /
        Telegram / WhatsApp-OTP sign-in). SameSite=None in production so
        the Vercel-hosted SPA can call the API cross-site. Primary
        mechanism for browser clients.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User-session Bearer fallback — `Authorization: Bearer <token>`
        accepted wherever cookieAuth is (requireUser checks the cookie
        first, then the header).
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Admin session — separate httpOnly cookie (8h) issued by POST
        /admin/login (and verify-2fa when TOTP is enabled).
    adminBearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin-session Bearer alternative — the token returned in the
        POST /admin/login JSON body, accepted wherever adminCookieAuth
        is (requireAdmin accepts either).
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"
        details:
          type: object
          additionalProperties: true
          description: |
            Optional structured context — zod `issues` on validation
            failures, `{field}` hints on flash-sale validation errors,
            `{skipped_duplicates}` on inventory uploads.
      required: [error, code]

    # Machine-readable error codes — mirrors the shared ErrorCode enum
    # in @workspace/error-codes (shared/error-codes/src/index.ts), the
    # single source both backend and frontend re-export. The copilot
    # preview/execution family (round-92 B3/F-05) closes the gap where
    # backend/src/routes/admin/copilot/* returned codes outside the enum.
    ErrorCode:
      type: string
      enum:
        - INVALID_DATA
        - INVALID_PHONE
        - INVALID_PASSWORD_LENGTH
        - INVALID_PASSWORD_WEAK
        - INVALID_OTP
        - INVALID_CREDENTIAL
        - UNAUTHORIZED
        - INVALID_TOKEN
        - SESSION_EXPIRED
        - ACCOUNT_LOCKED
        - ACCOUNT_NOT_FOUND
        - PHONE_ALREADY_REGISTERED
        - FEATURE_DISABLED
        - FORBIDDEN
        - INSUFFICIENT_PERMISSIONS
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - OUT_OF_STOCK
        - PRODUCT_UNAVAILABLE
        - INSUFFICIENT_BALANCE
        - INVALID_AMOUNT
        - TOPUP_LIMIT_EXCEEDED
        - ORDER_NOT_FOUND
        - ORDER_ALREADY_COMPLETED
        - ORDER_CANNOT_CANCEL
        - GOOGLE_TOKEN_INVALID
        - GOOGLE_VERIFICATION_FAILED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - RATE_LIMITED
        - IDEMPOTENCY_IN_FLIGHT
        - IDEMPOTENCY_KEY_REUSE
        - COPILOT_INVALID_INPUT
        - COPILOT_LLM_ERROR
        - COPILOT_NO_ADMIN_SESSION
        - COPILOT_BAD_METHOD
        - COPILOT_PREVIEW_NOT_FOUND
        - COPILOT_PREVIEW_CONSUMED
        - COPILOT_PREVIEW_EXPIRED
        - COPILOT_HANDOFF_REQUIRED
        - COPILOT_HIGH_RISK_DISABLED
        - COPILOT_STALE_RECORD
        - COPILOT_UNEXPECTED_STATE
        - COPILOT_EXECUTE_FAILED
        - COPILOT_NOT_HIGH_RISK
        - COPILOT_FIRST_CONFIRM_MISSING
        - COPILOT_COOLDOWN_NOT_ELAPSED
        - COPILOT_LLM_UNAVAILABLE
        - COPILOT_OUT_OF_SCOPE
        - COPILOT_INVALID_FLAGS
        # Round-93 (93-A8 F-8/F-15): two emitted codes that previously
        # lived outside the whole code contract — the copilot rate
        # limiter's 429 body and the CSRF gate's fail-closed
        # misconfiguration branch. The enum stays exhaustive over what
        # the backend actually emits.
        - COPILOT_RATE_LIMITED
        - CSRF_CONFIG
        # Round-94 (A5-05): the last two emitted-but-unenumerated codes —
        # the copilot phase-gate 503 (requireCopilotPhase) and the
        # outbound secret-scan 502 abort (ask/draft routes).
        - COPILOT_PHASE_DISABLED
        - COPILOT_SECRET_LEAK

    # 409 body returned by the idempotency middleware
    # (backend/src/middlewares/idempotency.ts): the same Idempotency-Key
    # re-sent with a different body (IDEMPOTENCY_KEY_REUSE) or the same
    # key while the original request is still executing
    # (IDEMPOTENCY_IN_FLIGHT). `message` is a legacy alias kept for older
    # clients — `error` is the canonical field.
    IdempotencyConflictResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Always false on these conflicts.
        error:
          type: string
        message:
          type: string
        code:
          type: string
          enum: [IDEMPOTENCY_IN_FLIGHT, IDEMPOTENCY_KEY_REUSE]
      required: [success, error, message, code]

    # 429 body from POST /wallet/topups — the standard {error, code}
    # envelope plus pending-count context (wallet.ts MAX_PENDING guard;
    # not produced by createErrorResponse).
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        display_name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photo_url:
          type: ["string", "null"]
        auth_provider:
          type: string
          description: |
            "firebase_phone" | "firebase_google" | "firebase" |
            "telegram" | "whatsapp_phone" | "legacy_password".
        has_google:
          type: boolean
        has_telegram:
          type: boolean
        has_firebase:
          type: boolean
        has_whatsapp:
          type: boolean
        last_auth_at:
          type: ["string", "null"]
        wallet_balance:
          type: number
        loyalty_points:
          type: integer
        loyalty_tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        lifetime_spend:
          type: number
        order_count:
          type: integer
          description: Completed-order count for this user.
        referral_code:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        [
          id,
          phone,
          wallet_balance,
          loyalty_points,
          loyalty_tier,
          lifetime_spend,
          order_count,
          created_at,
        ]

    # PATCH /admin/users/{id} body. At least one editable field must
    # be supplied ("no updates" → 400). Wallet fields route through
    # AdjustmentService (ledger-backed); wallet_adjustment (relative
    # delta) takes precedence when both wallet fields are present.
    UpdateAdminUserBody:
      type: object
      properties:
        wallet_adjustment:
          type: number
          description: |
            Relative delta applied to wallet_balance via AdjustmentService
            (transaction + wallet_ledger type=adjustment). Rejected with
            400 on non-finite/oversized values (INVALID_AMOUNT), zero
            delta, or a resulting negative balance.
        wallet_balance:
          type: number
          description: |
            Absolute balance set via AdjustmentService (same ledger path).
            Ignored when wallet_adjustment is also present.
        loyalty_points:
          type: integer
          minimum: 0
          maximum: 10000000
          description: |
            Direct-update path (no ledger). Integer in [0, 10,000,000] —
            the cap exists because points convert to wallet credit via
            /loyalty/convert-points (100 pts/LYD).
        loyalty_tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        note:
          type: string
          description: |
            Free-form note recorded on the wallet_ledger row when a
            wallet field is supplied (defaults to "Admin adjustment" /
            "Admin balance set").
      minProperties: 1

    # 200 body of PATCH /admin/users/{id} — the current post-edit state
    # of the mutated fields.
    AdminUserUpdateResult:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        wallet_balance:
          type: number
        loyalty_points:
          type: integer
        loyalty_tier:
          type: string
          enum: [bronze, silver, gold, platinum]
      required: [id, phone, wallet_balance, loyalty_points, loyalty_tier]

    # ── Admin flash sales ─────────────────────────────────────────────────
...